fix packet parsing
